+++ shotmanager/ui/sm_ui.py
@@ -676,6 +676,8 @@ class UAS_PT_ShotManager(Panel):
             iconCheckBoxes = "CHECKBOX_HLT" if not prefs.toggleShotsEnabledState else "CHECKBOX_DEHLT"
             subrowedit.operator("uas_shot_manager.enabledisableall", text="", icon=iconCheckBoxes, emboss=False)
 
+            subrowedit.prop(props, "useContinuousGPEditing", text="", icon="GREASEPENCIL")
+
             if props.display_editmode_in_properties:
                 subrowedit.prop(
                     props,